feat: [#1096] ban client IP when exceeds connection ID errors limit
The life demo tracker is receiving many UDP requests with a wrong conenctions IDs. Errors are logged (write disk) and that
decreases the tracker performance.

This counts errors and bans Ips after 10 errors for 2 minutes.

We use two levels of counters.

1. First level: A Counting Bloom Filter: fast and low memory consumption
   but innacurate (False Positives).
2. HashMap: Exact Counter for Ips.

CBFs are fast and use litle memory but they are also innaccurate. They
have False Positives meaning some IPs would be banned only becuase there
are bucket colissions (IPs sharing the same counter).

To avoid banning IPs incorrectly we decided to introduce a second
counter, which is a HashMap that counts error precisely. IPs are only
banned when this counter reaches the limit (over 10 errors).

We keep the CBF as a first level filter. It's a fast-check IP filter
without affecting tracker's performance. When the IP is banned according
to the first filter we double-check in the HashMap.

CBF is faster than checking always for banned IPs against the HashMap.

This solution should be good if the number of IPs is low. We have to
find another solution anyway for IPv6 where is cheaper to own a range of
IPs.