fix: [1097] do not compile slow doc tests
It takes 32 seconds to compile and run all doc tests inckuding the ones
that use the main tracker lib (each test compiles the main lib).

This change disabled the compilation and execution of those tests.

In terms of code coverage is not a problem becuase we have units test
for that functionality. It only affects to the documentation. However
there are too slow and they are even making crah some developers'
computers.

The good solution would be to extract those parts into new workspace
packages to avoid compiling the main library. A new issue has been
opened for it:

https://github.com/torrust/torrust-tracker/issues/1140