test: capture logs in tests
This feature will be used in the future to write assertions about logs.

It also changes when we show logs running tests. If you run tests with:

```console
cargo test
```

logs win't be showed. If you want to see logs you have to execute tests
with:

```console
cargo test -- --nocapture
```