fix: missing logs for HTTP requests
Some tracing log spans were missing. In adition to that we were using
new spans to log requests and responses instead of using the span
provided by the TraceLayer. That enables to join all the log events
related tho the same requests.

This fix is needed to contunie with this issue:

https://github.com/torrust/torrust-tracker/issues/1150

Becuase it allos to use a "request-id" header to identify logs.

We can write log assertions by mathich lines with the request-id used in
the request.

For example, it yo make this request:

```console
curl -H "x-request-id: YOUR_REQUEST_ID" http://0.0.0.0:1212/api/v1/stats?token=InvalidToken
```

This is the new output (which was missing before this change):

```output
2024-12-23T17:53:06.530704Z  INFO request{method=GET uri=/api/v1/stats?token=InvalidToken version=HTTP/1.1}: API: request method=GET uri=/api/v1/stats?token=InvalidToken request_id=YOUR_REQUEST_ID
2024-12-23T17:53:06.530777Z ERROR request{method=GET uri=/api/v1/stats?token=InvalidToken version=HTTP/1.1}: API: response latency_ms=0 status_code=500 Internal Server Error request_id=YOUR_REQUEST_ID
2024-12-23T17:53:06.530785Z ERROR request{method=GET uri=/api/v1/stats?token=InvalidToken version=HTTP/1.1}: API: response failed failure_classification=Status code: 500 Internal Server Error latency=0 ms
```

As you can see. now we have the "request_id=YOUR_REQUEST_ID" field which
can be used to identify the test that made the request.