feat: make ban service generic for all trackers
All UDP tracker will share the same service. In the future, the HTTP
trackers can also use it.

The service was not include inside the tracker (easy solution) becuase
the Tracker type is too big. It has became the app container. In fact,
we want to reduce it in the future by extracting the services outside of
the tracker: stats, whitelist, etc. Those services will be instantiate
independently in the future in the app bootstrap.