refactor: [#1184] inject stats event sender and repository into the core tracker
This is part of a big refactor. We are extracting responsabilities from
the tracker. The first step is to inject the service into the tracker
and later we will use the extracted services directly. Finally we will
removed the injected service from the tracker when it's not used anymore
via the tracker.